requirements.txt
pandas
keras
flask
gunicorn
numpy
tensorflow-macos
tensorflow-metal
sklearn